backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0

# Data science
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.1

# PDF export
reportlab>=4.0.0
